Fixed bug in mpi_exchanges related to mpi masking.
Needed to replace the if west_msg.and.north_msg logicals to decide
to exchange with the subdomain to the upper left with if p_NW>0.
This was never an issue, but now it's possible to have sub-domains
to your north and west, but not have one to your north-west.
This was a tricky one! (But obvious in hindsight, of course)