Significant changes to mpi as well as initialization
mpi now works with checking node numbers to exchange with instead
of 'west_exch' etc. Initialize vars, especially grid related variables
to be those in masked areas (using hmin etc). This way, we don't need
to use exchange functions to get workable values in the buffers of
variables in 'working' subdomains that are next to 'ignored' subdomains