Removed extra call to cstar_output in main, and added avg_time variables.
There was a redundant call to write cstar output in main, which is now removed.
Also added two variables, avg_begin_time and avg_end_time, that are written in the cstar
output when do_avg is true.  These variables let the user know when the averaging period
for each record begins and ends.