Single-file CDR forcing module.
This commit expands the ncforce type to include a "single-file" option,
where all of the data is read from one file and only on the head node. The
data is then broadcasted to the other MPI ranks. The filename no longer
needs to be listed in the Forcing section of the .in file, but it still needs
to be specified somewhere else and declared in the nc%single_filename field.