Removed normalization by Hz for cdr_prf.
In the CDR forcing module the vertical distribution of the forcing was
being weighted by a multiplication by Hz, which was making the distribution of
the tracer mass look weird when they were plotted. Removal of this
multiplication made the distribution much more sensible, as it was no longer
affected by cell thickness; now the amount of mass added to each cell is only
affected by its distance from the release location.