Fix for exact restarts involving rivers.
Rivers works by overwriting u and ubar at the end of the barotropic
and baroclinic steps. However, they are unique in that they occur at
interfaces between wet and dry points, meaning that they actually occur
where umask and vmask are zero. Previously, when the velocities would
be read from file during initialization, they would be zeroed out
wherever umask or vmask are zero, meaning that the river velocities were
removed. This was breaking exact restarts. This commit fixes that by
saving a "river mask" in the restart files, which are read in during
the restart phase and tell the model not to zero out the rivers.