Made background visc/diff additive with kpp.
It used to be mutually exclusive, meaning that the "vertical mixing"
values in the .in file were not even used whenever kpp is active (with no
warning). This PR makes it so that ROMS always reads "vertical mixing", and
its values become the global background that kpp is added on top of.