Fix for exact restarts with CDR forcing.
For parameterized forcing the module was calculating how to distribute
the forcing based on a vertical structure function that was calculated using
the variable z_r. This implicitly used the SSH at the time of calculation,
but since it is only calculated during model initialization this means that
the vertical structure would be slightly different after any restart. This
bugfix creates new depth/thickness variables that are calculated assuming zero
SSH, which means that the CDR vertical structures will be the same regardless
of when they are calculated.